Decided to make Terminal implement Closeable after all
This makes the code cleaner when trying to shutdown an application. Previously, we had to do instanceof to figure out if it was a JFrame that needed to be disposed unless the auto-close triggered. I think it's better this way, we'll just leave the implementations that don't care about close to do nothing.