Fix and Test for Issue #353 rebased to release/3.0
This fixes an exception thrown for trying to supply Window.HINTs to the MessageDialogBuilder.

Also, change MessageDialogBuilder.setExtraWindowHints to accept any
Collection<Window.HINT>, as there is really no need to confine to a Set.

Give Window.Hint a decent .toString() for debugging.

This fix is already merged to master, but might be good for release/3.0, too.