Small override to Window.pack() to allow fullscreen keyboard input
I added a small method to SwingTerminalFrame.java overriding Java AWT's Window.pack() method.

All that the override does is make a call to super.pack() before setting disposed to false.

This change was made because the DefaultTerminalFactory used to create a SwingTerminalFrame returns a displayable Frame, which does not allow for the use of the Frame.setUndecorated() method, which can be used to create a borderless, fullscreen window. This may seem irrelevant to keyboard input, but becomes a problem if one wants to use the SwingTerminalFrame.dispose() method. This method would previously set the private boolean disposed variable to false, which would result in the SwingTerminalFrame.pollInput() function erroneously ceasing to read keyboard input after the SwingTerminalFrame.pack() method was called afterwards.