add getColumn() method
I ended up needing to get the contents of a column as a list similar to getRow() pretty frequently and I assume others will as well so i added a method for that.  As far as I know table data is stored as a list of rows so this does require iterating over each row in the table at least the way I did it.  If there is a better way to do this please let me know.